refactor(cli-sub-agent)!: remove dev2merge subcommand, add generic plan run --issue
The `csa dev2merge` subcommand was a thin compiled alias for
`csa plan run --pattern dev2merge` whose only unique capability was
`--issue <N>` (fetch a GitHub issue body into the FEATURE_INPUT workflow
variable). In headless mode it paused at the manual mktsk handoff and
exited 0 -- a silent false success that produced a plan but never
implemented it. The architecturally-correct entry point is the
`/dev2merge` skill plus the pattern, so the compiled subcommand is a
category error.

Remove the subcommand entirely (pre-production: no deprecation alias) and
sink its useful `--issue <N>` capability into `csa plan run` as a generic
flag usable by any pattern that reads FEATURE_INPUT:

- Delete cli_dev2merge.rs and dev2merge_cmd.rs; drop the Commands::Dev2merge
  enum variant, its main.rs dispatch arm, the cli_plan.rs module include,
  and the sa_mode / executor_csa_guard / cli_review match arms.
- Add `--issue <N>` to `csa plan run`. When set, the top-level invocation
  fetches the issue body once (fail-fast on a bad number or auth error)
  and injects it as the FEATURE_INPUT variable; the resolved `--var` is
  forwarded to the daemon child in place of `--issue` so the child never
  re-fetches.
- `--issue` is mutually exclusive with an explicit `--var FEATURE_INPUT=...`
  and fails loudly if the fetch fails after the auth fallback.
- Move the GitHub issue fetch (GH_CONFIG_DIR auth + default-auth fallback)
  into the gh_env module beside resolve_gh_env/is_auth_error.
- Extract the plan-run journal, resume-context, and repo-fingerprint
  primitives into a new plan_cmd_journal module (re-exported from plan_cmd
  for the daemon dispatch and the in-module test submodules), keeping
  plan_cmd.rs within the per-file 800-line / 8000-token module budget.

Refs #1638
