+++ crates/cli-sub-agent/src/sa_mode_tests.rs
@@ -41,7 +41,7 @@ mod tests {
     }
 
     #[test]
-    fn sa_mode_parses_for_batch_plan_dev2merge_and_claude_sub_agent() {
+    fn sa_mode_parses_for_batch_plan_and_claude_sub_agent() {
         let batch_cli = Cli::try_parse_from(["csa", "batch", "task.toml", "--sa-mode", "true"])
             .expect("batch cli should parse");
         match batch_cli.command {
@@ -59,24 +59,36 @@ mod tests {
             _ => panic!("expected plan command"),
         }
 
-        let dev2merge_cli =
-            Cli::try_parse_from(["csa", "dev2merge", "--sa-mode", "true", "--issue", "1287"])
-                .expect("dev2merge cli should parse");
-        match dev2merge_cli.command {
-            Commands::Dev2merge(args) => {
-                assert_eq!(args.sa_mode, Some(true));
-                assert_eq!(args.issue, Some(1287));
-                assert_eq!(args.timeout, None);
-            }
-            _ => panic!("expected dev2merge command"),
-        }
-
         let claude_cli =
             Cli::try_parse_from(["csa", "claude-sub-agent", "--sa-mode", "true", "question"])
                 .expect("claude-sub-agent cli should parse");
         match claude_cli.command {
             Commands::ClaudeSubAgent(args) => assert_eq!(args.sa_mode, Some(true)),
             _ => panic!("expected claude-sub-agent command"),
+        }
+    }
+
+    #[test]
+    fn plan_run_parses_issue_flag() {
+        let plan_cli = Cli::try_parse_from([
+            "csa",
+            "plan",
+            "run",
+            "flow.toml",
+            "--sa-mode",
+            "true",
+            "--issue",
+            "1287",
+        ])
+        .expect("plan run --issue cli should parse");
+        match plan_cli.command {
+            Commands::Plan {
+                cmd: PlanCommands::Run { issue, sa_mode, .. },
+            } => {
+                assert_eq!(issue, Some(1287));
+                assert_eq!(sa_mode, Some(true));
+            }
+            _ => panic!("expected plan command"),
         }
     }
 
@@ -93,11 +105,14 @@ mod tests {
     }
 
     #[test]
-    fn top_level_help_lists_dev2merge_subcommand() {
+    fn top_level_help_omits_dev2merge_subcommand() {
+        // The `csa dev2merge` subcommand was removed in #1638; the dev2merge
+        // pipeline is reached via `csa plan run patterns/dev2merge/workflow.toml`
+        // and the `/dev2merge` skill. Guard against accidental reintroduction.
         let help = Cli::command().render_long_help().to_string();
         assert!(
-            help.contains("dev2merge"),
-            "top-level help should list dev2merge subcommand:\n{help}"
+            !help.contains("dev2merge"),
+            "top-level help must NOT list a dev2merge subcommand (removed in #1638):\n{help}"
         );
     }
 
@@ -149,7 +164,6 @@ mod tests {
             &["csa", "debate", "question"],
             &["csa", "batch", "task.toml"],
             &["csa", "plan", "run", "flow.toml"],
-            &["csa", "dev2merge"],
             &["csa", "claude-sub-agent", "question"],
         ];
